src/backend/requirements.txt
requests
pymupdf
openai
psycopg2-binary
python-dotenv
